requirements.txt
GDAL==2.2.2
lxml==4.3.0
matplotlib==3.0.0
numba==0.41.0
numpy==1.15.4
Pillow==5.3.0
scipy==1.1.0
utm==0.4.2
imageio==2.5.0
opencv-python==4.0.0
pyquaternion==0.9.5
pymap3d==1.7.15
python-dateutil==2.8.0

